+++ cmd/klew-desktop/frontend/wailsjs/go/main/App.d.ts
@@ -16,6 +16,8 @@ export function GetCluster():Promise<kube.ClusterState>;
 
 export function GetClusterStatus():Promise<kube.ClusterStatus>;
 
+export function GetKubectlInfo():Promise<kube.KubectlInfo>;
+
 export function GetObjectDetails(arg1:string,arg2:string,arg3:string):Promise<details.ObjectDetail>;
 
 export function GetResourceCatalog(arg1:main.CatalogOptions):Promise<model.ResourceCatalog>;
@@ -47,6 +49,8 @@ export function SelectNamespace(arg1:string):Promise<kube.ClusterState>;
 export function SetAutoRefresh(arg1:boolean):Promise<void>;
 
 export function SetKubeconfigPath(arg1:string):Promise<kube.ClusterState>;
+
+export function SetKubectlOptions(arg1:boolean,arg2:string,arg3:boolean):Promise<kube.KubectlInfo>;
 
 export function SetPollEverySec(arg1:number):Promise<void>;